Refactor data layer: Split EntityManager, add EntityDataManager and EntityAPIFactory
