+++ src/core/entity-api.ts
@@ -1,5 +1,6 @@
 import type { Context, EntityConfiguration } from '../types';
 import type { EntityManager } from './entity-manager';
+import type { EntityDataManager } from './entity-data-manager';
 import type { ValidationManager } from './validation-manager';
 import type { PermissionManager } from './permission-manager';
 import type { WorkflowManager } from './workflow-manager';
@@ -12,6 +13,7 @@ export class EntityAPI {
   constructor(
     private readonly entityName: string,
     private readonly entityManager: EntityManager,
+    private readonly entityDataManager: EntityDataManager,
     private readonly validationManager: ValidationManager,
     private readonly permissionManager: PermissionManager,
     private readonly workflowManager: WorkflowManager,
@@ -27,7 +29,7 @@ export class EntityAPI {
     await this.enforcePermission(entityConfig, 'create', contextWithTenant);
     await this.validateData(entityConfig, data, 'create');
 
-    const result = await this.entityManager.insertData(entityConfig, data, contextWithTenant);
+    const result = await this.entityDataManager.insertData(entityConfig, data, contextWithTenant);
 
     await this.workflowManager.executeWorkflows(entityConfig, 'create', null, result, contextWithTenant);
 
@@ -47,7 +49,7 @@ export class EntityAPI {
       operator: 'eq'
     }));
 
-    return this.entityManager.findData(entityConfig, conditions, {}, contextWithTenant);
+    return this.entityDataManager.findData(entityConfig, conditions, {}, contextWithTenant);
   }
 
   async update(id: string | number, data: Record<string, any>, context: Context = {}) {
@@ -57,8 +59,8 @@ export class EntityAPI {
     await this.enforcePermission(entityConfig, 'update', contextWithTenant);
     await this.validateData(entityConfig, data, 'update');
 
-    const oldData = await this.entityManager.findByIdData(entityConfig, id, contextWithTenant);
-    const result = await this.entityManager.updateData(entityConfig, id, data, contextWithTenant);
+    const oldData = await this.entityDataManager.findByIdData(entityConfig, id, contextWithTenant);
+    const result = await this.entityDataManager.updateData(entityConfig, id, data, contextWithTenant);
 
     await this.workflowManager.executeWorkflows(entityConfig, 'update', oldData, result, contextWithTenant);
 
@@ -71,8 +73,8 @@ export class EntityAPI {
 
     await this.enforcePermission(entityConfig, 'delete', contextWithTenant);
 
-    const oldData = await this.entityManager.findByIdData(entityConfig, id, contextWithTenant);
-    const result = await this.entityManager.deleteData(entityConfig, id, contextWithTenant);
+    const oldData = await this.entityDataManager.findByIdData(entityConfig, id, contextWithTenant);
+    const result = await this.entityDataManager.deleteData(entityConfig, id, contextWithTenant);
 
     await this.workflowManager.executeWorkflows(entityConfig, 'delete', oldData, null, contextWithTenant);
 
@@ -85,7 +87,7 @@ export class EntityAPI {
 
     await this.enforcePermission(entityConfig, 'read', contextWithTenant);
 
-    return this.entityManager.findByIdData(entityConfig, id, contextWithTenant);
+    return this.entityDataManager.findByIdData(entityConfig, id, contextWithTenant);
   }
 
   // ----- Schema and Configuration getters -----
