+++ src/schemakit.ts
@@ -1,10 +1,7 @@
 import { SchemaKitOptions } from './types';
 import { DatabaseManager, DatabaseConfig } from './database/database-manager';
 import { InstallManager } from './database/install-manager';
-import { EntityManager } from './core/entity-manager';
-import { ValidationManager } from './core/validation-manager';
-import { PermissionManager } from './core/permission-manager';
-import { WorkflowManager } from './core/workflow-manager';
+import { EntityAPIFactory } from './core/entity-api-factory';
 import { SchemaKitError } from './errors';
 
 export class SchemaKit {
@@ -12,10 +9,7 @@ export class SchemaKit {
   private readonly databaseManager: DatabaseManager;
 
   private installManager?: InstallManager;
-  private entityManager?: EntityManager;
-  private validationManager?: ValidationManager;
-  private permissionManager?: PermissionManager;
-  private workflowManager?: WorkflowManager;
+  private entityAPIFactory?: EntityAPIFactory;
 
   constructor(options: SchemaKitOptions = {}) {
     this.options = options;
@@ -37,11 +31,8 @@ export class SchemaKit {
     try {
       await this.databaseManager.connect();
 
-      // Initialize managers with DatabaseManager
-      this.entityManager = new EntityManager(this.databaseManager);
-      this.validationManager = new ValidationManager();
-      this.permissionManager = new PermissionManager(this.databaseManager.getAdapter());
-      this.workflowManager = new WorkflowManager(this.databaseManager.getAdapter());
+      // Initialize EntityAPIFactory with DatabaseManager (handles all manager creation)
+      this.entityAPIFactory = new EntityAPIFactory(this.databaseManager);
 
       // Initialize system tables
       this.installManager = new InstallManager(this.databaseManager.getAdapter());
@@ -60,10 +51,10 @@ export class SchemaKit {
    * @param tenantId Tenant identifier (defaults to 'default')
    */
   entity(name: string, tenantId: string = 'default') {
-    if (!this.entityManager) {
+    if (!this.entityAPIFactory) {
       throw new SchemaKitError('SchemaKit is not initialized. Call `initialize()` first.');
     }
-    return this.entityManager.createEntityAPI(name, tenantId);
+    return this.entityAPIFactory.createEntityAPI(name, tenantId);
   }
 
   /**
@@ -71,6 +62,16 @@ export class SchemaKit {
    */
   getDatabase(): DatabaseManager {
     return this.databaseManager;
+  }
+
+  /**
+   * Access entity manager for configuration management
+   */
+  getEntityManager() {
+    if (!this.entityAPIFactory) {
+      throw new SchemaKitError('SchemaKit is not initialized. Call `initialize()` first.');
+    }
+    return this.entityAPIFactory.getEntityManager();
   }
 
   /**
